Updating command handler, the comment.

+++ filter/f_base.cpp
@@ -127,17 +127,17 @@ bool f_base::s_fpar::get(char * valstr, size_t sz){
 void f_base::s_fpar::get_info(s_cmd & cmd){
 	snprintf(cmd.get_ret_str(), CMD_LEN, "%s: %s", name, explanation);
 	if(type == ENUM){
-		int slen = strlen(cmd.get_ret_str());
+		int slen = (int) strlen(cmd.get_ret_str());
 		char * ptr = cmd.get_ret_str() + slen;
 		*ptr = ' ';
 		snprintf(ptr, CMD_LEN - slen, " Vals in {");
 		ptr = cmd.get_ret_str();
-		slen = strlen(ptr);
+		slen = (int) strlen(ptr);
 		for(int i = 0; i < len && slen < len; i++){
 			snprintf(ptr + len, CMD_LEN - slen, "%s,", str_enum[i]);
-			slen += strlen(str_enum[i] + 1);
+			slen += (int) strlen(str_enum[i] + 1);
 		}
-		slen = strlen(cmd.get_ret_str());
+		slen = (int) strlen(cmd.get_ret_str());
 		ptr[slen-1] = '}';
 	}
 }
